backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.6.1
pydantic-settings==2.0.3
python-dotenv==1.0.1
email-validator==2.1.0.post1
fastapi-mail==1.4.1
bcrypt==4.1.2
